Refactor package `env` to `version`
