+++ tools/goctl/goctl.go
@@ -24,6 +24,7 @@ import (
 	"github.com/tal-tech/go-zero/tools/goctl/configgen"
 	"github.com/tal-tech/go-zero/tools/goctl/docker"
 	"github.com/tal-tech/go-zero/tools/goctl/internal/errorx"
+	"github.com/tal-tech/go-zero/tools/goctl/internal/version"
 	"github.com/tal-tech/go-zero/tools/goctl/kube"
 	"github.com/tal-tech/go-zero/tools/goctl/model/mongo"
 	model "github.com/tal-tech/go-zero/tools/goctl/model/sql/command"
@@ -34,8 +35,7 @@ import (
 )
 
 var (
-	buildVersion = "1.1.11-beta-2"
-	commands     = []cli.Command{
+	commands = []cli.Command{
 		{
 			Name:   "upgrade",
 			Usage:  "upgrade goctl to latest version",
@@ -650,14 +650,10 @@ func main() {
 
 	app := cli.NewApp()
 	app.Usage = "a cli tool to generate code"
-	app.Version = fmt.Sprintf("%s %s/%s", buildVersion, runtime.GOOS, runtime.GOARCH)
+	app.Version = fmt.Sprintf("%s %s/%s", version.BuildVersion, runtime.GOOS, runtime.GOARCH)
 	app.Commands = commands
 	// cli already print error messages
 	if err := app.Run(os.Args); err != nil {
 		fmt.Println(aurora.Red(errorx.Wrap(err).Error()))
 	}
-}
-
-func init() {
-	_ = os.Setenv("GOCTL_VERSION", buildVersion)
 }
