+++ tools/goctl/internal/version/version_test.go
@@ -1,4 +1,4 @@
-package env
+package version
 
 import (
 	"testing"
